backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pymongo[srv]>=4.6.0
pydantic>=2.6.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
